# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Prueba Tecnica Spot2",
        "description": "",
        "version": "1.0.0"
    },
    "paths": {
        "/price-m2/zip-codes/{zipCode}/aggregate/{type}": {
            "get": {
                "tags": [
                    "Precio por M2"
                ],
                "description": "Permite obtener el precio por m2",
                "operationId": "3f7bae6cfa3d06e7df0971a2cfd3bab9",
                "parameters": [
                    {
                        "name": "zipCode",
                        "in": "path",
                        "description": "Zip code to fetch data for (string)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "path",
                        "description": "Type of aggregation (min, max, avg)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "min",
                                "max",
                                "avg"
                            ]
                        }
                    },
                    {
                        "name": "uso_construccion",
                        "in": "query",
                        "description": "Usage construction parameter (1 to 7)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 7,
                            "minimum": 1,
                            "example": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "payload": {
                                            "properties": {
                                                "type": {
                                                    "type": "string",
                                                    "example": "avg"
                                                },
                                                "price_unit": {
                                                    "type": "integer",
                                                    "example": 1420
                                                },
                                                "price_unit_construction": {
                                                    "type": "integer",
                                                    "example": 3120
                                                },
                                                "elements": {
                                                    "type": "integer",
                                                    "example": 100
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No results found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Sin resultados para mostrar."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity: Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "status": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "error": {
                                                    "type": "string",
                                                    "example": "El uso_construccion es inválido."
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "status": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "error": {
                                                    "type": "string",
                                                    "example": "El aggregate es inválido."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Precio por M2",
            "description": "Precio por M2"
        }
    ]
}